+++ src/main/java/blackjack/domain/result/GameResultType.java
@@ -22,7 +22,6 @@ public enum GameResultType {
                 .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 비교 값입니다."));
     }
 
-
     public GameResultType getOppositeType() {
         if (this.equals(WIN)) {
             return LOSE;
